Split callbacks in individual files + add a  property to Callback for easy trainer instance access
